requirements.txt
boto3>=1.16,<2
slackclient>=2.9,<3
